Add `--relaxed` option to `check-conda-env`

+++ calkit/conda.py
@@ -4,8 +4,16 @@ import json
 import os
 import subprocess
 
+from pydantic import BaseModel
+
 import calkit
 from calkit import ryaml
+
+
+class EnvCheckResult(BaseModel):
+    env_exists: bool | None = None
+    env_needs_export: bool | None = None
+    env_needs_rebuild: bool | None = None
 
 
 def check_env(
@@ -13,18 +21,23 @@ def check_env(
     use_mamba=True,
     log_func=None,
     output_fpath: str = None,
-):
+    relaxed: bool = False,
+) -> EnvCheckResult:
     """Check that a conda environment matches its spec.
 
     If it doesn't match, recreate it.
 
     Note that this only works with exact or no version specification.
     Using greater than and less than operators is not supported.
+
+    If ``relaxed`` is enabled, dependencies can exist in either the conda or
+    pip category.
     """
     conda = "mamba" if use_mamba else "conda"
     if log_func is None:
         log_func = calkit.logger.info
     log_func(f"Checking conda env defined in {env_fpath}")
+    res = EnvCheckResult()
     envs = json.loads(
         subprocess.check_output([conda, "env", "list", "--json"]).decode()
     )["envs"]
@@ -44,11 +57,13 @@ def check_env(
     # Check if env even exists
     if env_name not in existing_env_names:
         log_func(f"Environment {env_name} doesn't exist; creating")
+        res.env_exists = False
         # Environment doesn't exist, so create it
         subprocess.check_call([conda, "env", "create", "-y", "-f", env_fpath])
         env_needs_rebuild = False
         env_needs_export = True
     else:
+        res.env_exists = True
         env_needs_export = False
         # Environment does exist, so check it
         if os.path.isfile(env_check_fpath):
@@ -67,6 +82,7 @@ def check_env(
         else:
             env_needs_export = True
         if env_needs_export:
+            res.env_needs_export = True
             log_func(f"Exporting existing env to {env_check_fpath}")
             env_check = json.loads(
                 subprocess.check_output(
@@ -100,6 +116,12 @@ def check_env(
         else:
             required_conda_deps = env_spec["dependencies"]
             required_pip_deps = []
+        if relaxed:
+            log_func("Running in relaxed mode; combining pip and conda deps")
+            for dep in existing_pip_deps:
+                existing_conda_deps.append(dep.replace("==", "="))
+            for dep in required_pip_deps:
+                required_conda_deps.append(dep.replace("==", "="))
         log_func("Checking conda dependencies")
         for dep in required_conda_deps:
             dep_split = dep.split("=")
@@ -121,7 +143,7 @@ def check_env(
                     log_func(f"Found missing dependency: {dep}")
                     env_needs_rebuild = True
                     break
-        if not env_needs_rebuild:
+        if not env_needs_rebuild and not relaxed:
             log_func("Checking pip dependencies")
             for dep in required_pip_deps:
                 dep_split = dep.split("==")
@@ -142,11 +164,13 @@ def check_env(
                         env_needs_rebuild = True
                         break
     if env_needs_rebuild:
+        res.env_needs_rebuild = True
         log_func(f"Rebuilding {env_name} since it does not match spec")
         subprocess.check_call([conda, "env", "create", "-y", "-f", env_fpath])
         env_needs_export = True
     else:
         log_func(f"Environment {env_name} matches spec")
+        res.env_needs_rebuild = False
     # If the env was rebuilt, export the env check
     if env_needs_export:
         log_func(f"Exporting existing env to {env_check_fpath}")
@@ -171,8 +195,14 @@ def check_env(
     if output_fpath is None:
         fname, ext = os.path.splitext(env_fpath)
         output_fpath = fname + "-lock" + ext
-    log_func(f"Exporting lock file to {output_fpath}")
-    with open(output_fpath, "w") as f:
-        _ = env_check.pop("mtime")
-        _ = env_check.pop("prefix")
-        ryaml.dump(env_check, f)
+    if (
+        not res.env_exists
+        or res.env_needs_rebuild
+        or not os.path.isfile(output_fpath)
+    ):
+        log_func(f"Exporting lock file to {output_fpath}")
+        with open(output_fpath, "w") as f:
+            _ = env_check.pop("mtime")
+            _ = env_check.pop("prefix")
+            ryaml.dump(env_check, f)
+    return res
